requirements.txt

Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
email-validator==2.1.0
python-dotenv==1.0.0
